refactor server moments / routes
